validacion delete

+++ src/main/java/com/heroes/service/HeroService.java
@@ -3,7 +3,9 @@ package com.heroes.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.heroes.model.Hero;
 import com.heroes.repository.HeroRepository;
@@ -19,7 +21,7 @@ public class HeroService {
 	}
 
 	public Hero getHeroById(long idHero) {
-		return heroRepository.findById(idHero).get();
+		return heroRepository.findById(idHero).orElse(null);
 	}
 
 	public Hero saveHero(Hero hero) {
@@ -30,11 +32,19 @@ public class HeroService {
 		return heroRepository.save(hero);
 	}
 
-	public void deleteHero(long idHero) {
+	public void deleteHero(long idHero) throws ResponseStatusException, Exception {
 		try {
+			Hero hero = getHeroById(idHero);
+
+			if (hero == null) {
+				throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Hero not found");
+			}
+
 			heroRepository.deleteById(idHero);
+		} catch (ResponseStatusException e) {
+			throw new ResponseStatusException(e.getStatus(), e.getReason());
 		} catch (Exception e) {
-			System.out.println("Exception: " + e.getMessage());
+			throw new Exception(e);
 		}
 	}
